nostr: split Double Ratchet modules
Separate session and crypto logic, invite and wire handling, tests, and interoperability fixtures without changing the public API or behavior.
