+++ nostr/src/nips/nip117/fixtures/typescript-first-event.json
@@ -1,0 +1,1 @@
+{"id":"cbf5c325bb54055050f148d9ae1a7197918142c6df76968ace985f0028a08bc8","pubkey":"4f355bdcb7cc0af728ef3cceb9615d90684bb5b2ca5f859ab0f0b704075871aa","created_at":1704067200,"kind":1060,"tags":[["header","Ai3RFOiEN/rJ+OO62lOm6Vbbzgl9vGVEhL0RSY7XHB2tj92twctf0ngH5fiRlu/sxNIvoggSrV0Oa1kMrJEw1rLrHhxtkEHAphpQOHgVUk0F9HSzRQAUmXuMEL85qgLtGQI9EuPz/lGSRi7sbocQWRxDe3j8VLFQkeyR6zPsNr2eG5EbPXyVD3+0rQMrHjf90dfbDPVr0Ts8uSf3XBd1Pba1FMyHt5KXtVS1XDDV1Q4qO5sijahDUquEQaNIqZSlc5AB"]],"content":"AvN/9kKCbZrDGZMFBke+mBoBCMcrS68GqusV895/yQPb9gcr4na152aO2ncVwe9MF0J+qMLZpidBg6JfzAw/mo2WlYLooZExNs6WAoLhEeNQ37BAhC9yepcnfCbum4XrckwkKUYDIhzpJjgCJbEFKp6sQr+cKKsUEBQM1aeLLAzzo4gB+FXbKZbYT7HwVX8hseutgDB5vsubHPCSA5lU4nf/mAEUapvEz2qJaS1ya0dZ2H1KZ/Ejcbo1oj3/WmjmgnDNXMCBHfo0p7G4+fGtaF5oyOsvciWiCOMY4aIzxAW3+5zPeRczVq6lbHo4tdhv3nAtetudSL1MRo8aoTvwTAm79zAcHNspdaHmVpix7JMjKvFv+XmkLn8d0OrQJ215uQSbF+9DnX1RiwiGdSBd7PsYCEqVlwVHfllor4PH4OTQsOQ=","sig":"9e75eb412702df314a9aa298e2841384493dc2cc76775c7879cc914faa4b279ca84b2d7ef436c00419c09981f9d7c6e7d33278bd91d0dd1b7e20cb706e706b93"}
