+++ nostr/src/nips/nip118/fixtures/typescript_invite_response.json
@@ -1,0 +1,9 @@
+{
+  "kind": 1059,
+  "pubkey": "d1d530e406de1b7cba221208598c8812c8b3e660615ceee1778fefbecf792e28",
+  "content": "Am9r/nq63KpvIw/YlsF3yMDA+9WIL9K6bvgE18xs9r4UFrJFgAMqY4DcDum9pzxMvj2hQ0zqiOlzu5c9/6udHfC/i+PCPsNdg6hQihqYRE0nQAXaLvxZAOWAeV+Gyc4dlrPLgzK4S8PJOOeC0yrTtTYAO/t5+HvOWQTqPUz+0GsEWXaUDKZ8dA9k5vX470QHvtOncWdkAyBRGdAptkyp8Q1KOjfh9yqMMRZRqXjenV5Kyj3uZOrGLndpsxtQZZoxCf3ofCIwPOR6nM0L4bopAHG/jlYcF6reXthTeS5B3usjE0oDooLu4/RY+xwBFiDSXC3AffbJMF+uaiFjSuERDIILe74XHUIDy7GcvhbFwLmYRCXcc7ocx7pLIZaP19suaBo2m7fxewQSJDStyhY2p3DS0ItrbRizMwwry0W/pUSOq65/l05I6+XKT2ZzJYIz1T67Dy74okTVqvyK3ITuNdFcqhRvcUsueqzia89Yfb5gUvcqUd16qhuPin+cA/yV5JApqT+eRqpPwpuAnc1SdeaedtsSsWNcxqtgwH0po6ydz3joZ+A2trWS4ubesK+p8BPFuCbAqcDBhcU/A1VVC8u6CUSwxfrklJAUIM8Xcb5R66zJl/TEvw+2MTwZS/f62dFGjv+502W/q9LogvNsBlAmsvG2/HD4FN/5SAIngD/ewg5DkI/TfReblveMmIVdG7ARmrI6HvvIpZyT944WfdChv2fY+B2vRUDCBNWSgCU3yajZxow9A1cDo0OOhYgnBtVlett0TjwYTpnfWR6vvUIt4zqwZ6JF2P/eMFo3C6qffqY0V88PjA5vfkg5GZsumI4Kv0Xio7RNXH8cj3y5KxLqzMQH//tej8rW5KLBLCtGV7Yu4rgKgYvdpngxHpXJkd5Gj0yfzG8C4tlblOmY9/jkeaeWw8sG37yQjosW/cDyVtxCjrdnjaWg1ehlcrYeG4w+8RYXZTdiHsQL75fIQ6xR4IedQqbk3MN7bl89EcsbL7CbbiaMAAbJG5jwtRQAzs7GW4FjjYlxH+7xDy/dFGc+mO4OLtX9ADBM4gSEialH2IomiM5EPXRug5I2ojwUsbpc+30yp3vO0jijrl73mcsAbJSB5yvD/zDrAI4gT2vVcwofQbi7pdTfhGe1Qe+yH7iPx7i8/FdP9GFlOsx9tMlKcKxrlRBKLUBmc1WPVBjwMdCjSSh9ib1cm0IZSlckdPaJA8zA1IW6/OwOog/oUjKMVIkjRlaEB/VMIA2VBnOVUyqPKXG4DG8oNyf8eJyaICaj",
+  "created_at": 1782876705,
+  "tags": [["p", "466d7fcae563e5cb09a0d1870bb580344804617879a14949cf22285f1bae3f27"]],
+  "id": "195a235d91ab72ebb7f82fad9a11f5439c5b920f83e3629e3733a7528e14e7a8",
+  "sig": "7b7a3fa68cc648deedfa178f5bb576e5de3ea8491717c2d42558caad45ad95e726434b96fe9e801272a68ec2ff55db0d68a3afd7430727099d35bdf2ff6b2054"
+}
